requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
lxml==5.2.2
requests==2.32.3
urllib3==2.2.2
pyyaml==6.0.2
